Optimized image read/write
Made all the images in "capture" store as PIL objects as opposed to jpgs. This means that the images are not stored on the disk until required(ie. writing a new control or saving the results) Instead they are stored on ram, as such the decision time was cut in half and capture time was slashed by a third.